# openapi.yaml
openapi: 3.0.3
info:
  title: Finanzas Backend API
  description: Personal finance ETL + API
  version: 0.3.0
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /etl/process/{file_id}:
    post:
      tags: [etl]
      summary: Run ETL pipeline for a source file
      description: |
        Parses, normalizes, categorizes, and persists transactions for the given
        source file. The bank/parser is resolved automatically from the account
        associated with the file.
        Returns immediately if the file was already processed.
      operationId: processEtl
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: ETL result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ETLProcessResponse'
        '404':
          description: Source file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '500':
          description: ETL processing failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
  /etl/reset/{file_id}:
    delete:
      tags: [etl]
      summary: Reset ETL for a source file
      description: |
        Deletes all transactions and raw rows associated with the file and
        resets its `parse_status` back to `UPLOADED`.
      operationId: resetEtl
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: Reset result
          content:
            application/json:
              schema:
                type: object
                properties:
                  file_id:
                    type: string
                    format: uuid
                  deleted_transactions:
                    type: integer
                  deleted_raw_rows:
                    type: integer
                  status:
                    type: string
                    example: UPLOADED
        '404':
          description: Source file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'

  # ──────────────────────────────────────────────
  # TRANSACTIONS
  # ──────────────────────────────────────────────
...
  /dashboard/kpis:
    get:
      tags: [dashboard]
      summary: KPI summary
      description: |
        Returns aggregate KPIs for the filtered set of transactions.
        All amounts are in COP (USD transactions converted at 4000 COP/USD).
        Results are scoped to the current user's accounts (admin sees all).

        Buckets are determined by transaction category:
        - `total_ingresos` — transactions categorized as `INGRESO` (money received into savings)
        - `total_gastos` — negative-amount transactions excluding `PAGO`, `INVERSION`, `INGRESO`, `MOVIMIENTO_ENTRE_BANCOS`
        - `total_pagos` — transactions categorized as `PAGO` (credit card payments from savings)
        - `total_inversiones` — transactions categorized as `INVERSION` (fund contributions)
        - `net` — `total_ingresos - total_gastos - total_inversiones` (`total_pagos` excluded to avoid double-counting CC expenses)
        - `avg_daily_spend` — `total_gastos` averaged over the number of distinct days in the result set
      operationId: getDashboardKpis
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OwnerFilter'
        - $ref: '#/components/parameters/AccountIdFilter'
        - $ref: '#/components/parameters/DateFromFilter'
        - $ref: '#/components/parameters/DateToFilter'
        - $ref: '#/components/parameters/CategoryFilter'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KPIResponse'
  /dashboard/histogram:
    get:
      tags: [dashboard]
      summary: Weekly expense histogram
      description: |
        Returns one data point per ISO week with aggregated `total_spent` in COP.
        Sorted ascending by week. PAGO transactions are always excluded.
        Results are scoped to the current user's accounts (admin sees all).
      operationId: getDashboardHistogram
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OwnerFilter'
        - $ref: '#/components/parameters/AccountIdFilter'
        - $ref: '#/components/parameters/DateFromFilter'
        - $ref: '#/components/parameters/DateToFilter'
        - $ref: '#/components/parameters/CategoryFilter'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HistogramPoint'
  /dashboard/by-category:
    get:
      tags: [dashboard]
      summary: Net spend breakdown by category
      description: |
        Groups expense transactions (amount < 0) by category and returns the absolute total per category.
        Categories `PAGO`, `INGRESO` and `INVERSION` are excluded (use `/kpis` for those).
        Only negative-amount transactions are included, so income entries that happen to share
        a category label do not distort the breakdown.
        Transactions without a category appear as `"SIN_CATEGORIZAR"`.
        Sorted by total descending. Results scoped to the current user (admin sees all).
      operationId: getDashboardByCategory
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OwnerFilter'
        - $ref: '#/components/parameters/AccountIdFilter'
        - $ref: '#/components/parameters/DateFromFilter'
        - $ref: '#/components/parameters/DateToFilter'
        - $ref: '#/components/parameters/CategoryFilter'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryBreakdownItem'
  /dashboard/top-transactions:
    get:
      tags: [dashboard]
      summary: Top N largest expenses
      description: |
        Returns the largest expenses (`amount < 0`) ordered by amount ascending
        (most negative first). All amounts are returned as positive values in COP.
        Results are scoped to the current user's accounts (admin sees all).
      operationId: getDashboardTopTransactions
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OwnerFilter'
        - $ref: '#/components/parameters/AccountIdFilter'
        - $ref: '#/components/parameters/DateFromFilter'
        - $ref: '#/components/parameters/DateToFilter'
        - $ref: '#/components/parameters/CategoryFilter'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
          description: Number of transactions to return
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TopTransactionItem'

# ──────────────────────────────────────────────
# COMPONENTS
# ──────────────────────────────────────────────
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    FileId:
      name: file_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Source file UUID
...
    OwnerFilter:
      name: owner
      in: query
      required: false
      schema:
        type: string
      description: Filter by owner label (e.g. "PACHO", "LU")
    AccountIdFilter:
      name: account_id
      in: query
      required: false
      schema:
        type: string
        format: uuid
      description: Filter by account UUID
    DateFromFilter:
      name: date_from
      in: query
      required: false
      schema:
        type: string
        format: date
        example: "2026-01-01"
      description: Start date (inclusive) in `YYYY-MM-DD` format
    DateToFilter:
      name: date_to
      in: query
      required: false
      schema:
        type: string
        format: date
        example: "2026-03-31"
      description: End date (inclusive) in `YYYY-MM-DD` format
    CategoryFilter:
      name: category
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/Category'
      description: Filter by transaction category
  schemas:
...
    BankEnum:
      type: string
      enum: [BANCOLOMBIA, RAPPI, FALABELLA, NEQUI]
    AccountTypeEnum:
      type: string
      enum: [CREDITO, DEBITO, AHORROS]
    Category:
      type: string
      enum:
        - HOGAR
        - DOMICILIOS
        - CARRO
        - TRANSPORTE
        - OCIO
        - RESTAURANTES
        - ROPA
        - SALUD
        - PRESTACIONES
        - REGALOS
        - EDUCACION
        - TRABAJO
        - COBRO_BANCARIO
        - PAGO
        - PLATAFORMAS
        - INGRESO
        - INVERSION
        - MOVIMIENTO_ENTRE_BANCOS
        - OTROS
...
    ETLProcessResponse:
      type: object
      properties:
        parsed_rows:
          type: integer
        inserted_transactions:
          type: integer
        duplicates_skipped:
          type: integer
        status:
          type: string
          enum: [PROCESSED, ALREADY_PROCESSED, FAILED]

    # ── Transactions ───────────────────────────
...
    KPIResponse:
      type: object
      required:
        - total_ingresos
        - total_gastos
        - total_pagos
        - total_inversiones
        - net
        - transaction_count
      properties:
        total_ingresos:
          type: number
          description: |
            Sum of positive-amount transactions in COP. Includes:
            - INGRESO-category transactions (money received into savings account)
            - Credits/refunds within spending categories (e.g. an OCIO refund)
        total_gastos:
          type: number
          description: Absolute sum of expense transactions in COP, excluding PAGO, INVERSION, INGRESO and MOVIMIENTO_ENTRE_BANCOS
        total_pagos:
          type: number
          description: Absolute sum of PAGO-category transactions in COP (credit card payments from savings)
        total_inversiones:
          type: number
          description: Absolute sum of INVERSION-category transactions in COP (contributions to investment funds)
        net:
          type: number
          description: |
            total_ingresos - total_gastos - total_inversiones.
            Note: total_pagos is excluded from net because credit card payments settle
            debt already counted as gastos in the CC accounts — subtracting them again
            would double-count those expenses.
        transaction_count:
          type: integer
        avg_daily_spend:
          type: number
          nullable: true
          description: total_gastos averaged over distinct days in the result set. Null when no data.
    HistogramPoint:
      type: object
      properties:
        week:
          type: string
          example: "2025-01-13"
          description: Start date (Monday) of the week in YYYY-MM-DD format
        total_spent:
          type: number
          description: Absolute sum of expenses for the week in COP
    CategoryBreakdownItem:
      type: object
      properties:
        category:
          type: string
          nullable: true
          description: Category name or "SIN_CATEGORIZAR" for uncategorized transactions
          example: RESTAURANTES
        total:
          type: number
          description: Total expenses for this category in COP (positive value)
        percentage:
          type: number
          description: Percentage of grand total (0–100)
          example: 23.45
        count:
          type: integer
          description: Number of transactions in this category
    TopTransactionItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        posted_at:
          type: string
          format: date
        description_clean:
          type: string
        amount:
          type: number
          description: Expense amount in COP, always positive (absolute value)
        category:
          type: string
          nullable: true
        merchant_guess:
          type: string
          nullable: true
        account_name:
          type: string
        bank_name:
          $ref: '#/components/schemas/BankEnum'
        owner:
          type: string
          description: Free-form owner label on the account
        account_type:
          $ref: '#/components/schemas/AccountTypeEnum'

    # ── Common ─────────────────────────────────
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string